+++ jdplus-main-base/jdplus-sa-base-parent/jdplus-sa-base-core/src/main/java/jdplus/sa/base/core/diagnostics/ResidualSeasonalityTests.java
@@ -96,7 +96,7 @@ public class ResidualSeasonalityTests {
         return test;
     }
 
-    public StatisticalTest kruskallWallisTest() {
+    public StatisticalTest kruskalWallisTest() {
         StatisticalTest test = kruskalWallisTest;
         if (test == null) {
             synchronized (this) {